+++ pyClarion/base/realizers.py
@@ -573,6 +573,26 @@ class ContainerConstruct(ConstructRealizer[It, Ot]):
 
     ctype: ClassVar[ConstructType] = ConstructType.container_construct
 
+    def __init__(
+        self, 
+        name: Hashable, 
+        matches: MatchSpec = None,
+        shared: Dict = None,
+        updaters: UpdaterArg[Any] = None,
+    ) -> None:
+        """
+        Initialize a new container realizer.
+        """
+
+        super().__init__(name=name, matches=matches, updaters=updaters)
+        # shared is a simple dict for storing/housing resources shared among 
+        # different components of the container including updaters and other 
+        # realizers. There is no sophisticated semantics, just a convenient 
+        # handle for storing handles to datastructures such as chunk databases 
+        # and bla information. It is the user's responsibility to make sure 
+        # shared resources are shared and used as intended.
+        self.shared = shared if shared is not None else dict()
+
     def __contains__(self, key: ConstructSymbol) -> bool:
 
         try:
@@ -786,10 +806,13 @@ class Subsystem(ContainerConstruct[ActivationPacket, SubsystemPacket]):
         name: Hashable, 
         matches: MatchSpec = None,
         proc: Callable[['Subsystem', Optional[Dict]], None] = None,
+        shared: Dict = None,
         updaters: UpdaterArg['Subsystem'] = None
     ) -> None:
 
-        super().__init__(name=name, matches=matches, updaters=updaters)
+        super().__init__(
+            name=name, matches=matches, shared=shared, updaters=updaters
+        )
         self.proc = proc
         self._features: Dict[ConstructSymbol, Node] = {}
         self._chunks: Dict[ConstructSymbol, Node] = {}
@@ -947,10 +970,13 @@ class Agent(ContainerConstruct[None, None]):
         self, 
         name: Hashable, 
         matches: MatchSpec = None, 
+        shared: Dict = None,
         updaters: UpdaterArg['Agent'] = None
     ) -> None:
 
-        super().__init__(name=name, matches=matches, updaters=updaters)
+        super().__init__(
+            name=name, matches=matches, shared=shared, updaters=updaters
+        )
         self._buffers: Dict[ConstructSymbol, Buffer] = {}
         self._subsystems: Dict[ConstructSymbol, Subsystem] = {}