+++ changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- attribute `shared` for container construct realizers. This is a simple dict for storing datastructures shared by multiple components of the parent realizer (e.g, chunk database may be shared by updaters).
 - `SubsystemPacket` class for reporting subsystem states.
 - Added `output_value` attribute to node realizers for easy inspection of current activation.
 - Added `ConstructType.from_str()`.